+++ CHANGELOG.md
@@ -15,6 +15,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - Approval process always allowing photo usage regardless of choice
 - [Accessibility] Announce title changes with aria-live
 
+### Removed
+
+- Feedback link into user survey
+
 ## [1.3.1] - 2021-02-18
 
 ### Fixed